requirements.txt
pybit
python-dotenv
pytest
pytest-cov
backtrader
yfinance
